+++ plugins/amaley-core/includes/class-amaley-core.php
@@ -31,6 +31,7 @@ class Amaley_Core {
         new Amaley_Core_Cluster_Cards();
         new Amaley_Core_SHG_Cards();
         new Amaley_Core_Member_Cards();
+        new Amaley_Core_Product_Origin_Panel();
 
         add_action( 'admin_enqueue_scripts', array( $this, 'enqueue_admin_assets' ) );
         add_action( 'admin_notices', array( $this, 'admin_dependency_notice' ) );